refactor: 운동 신청 검증 로직 이동
